FEAT: checks for presence of same user or admin.

+++ src/controllers/user.js
@@ -1,7 +1,10 @@
-const { createUserDb, getUsersDb, deleteUserDb } = require("../domains/user.js");
+const {
+    createUserDb,
+    getUsersDb,
+    deleteUserDb,
+} = require("../domains/user.js");
 const jwt = require("jsonwebtoken");
-const secret = process.env.JWT_SECRET
-
+const secret = process.env.JWT_SECRET;
 
 const createUser = async (req, res) => {
     const { username, password, role } = req.body;
@@ -11,30 +14,32 @@ const createUser = async (req, res) => {
             error: "Missing fields in request body",
         });
     }
-        const createdUser = await createUserDb(username, password, role);
+    const createdUser = await createUserDb(username, password, role);
 
-        const token = jwt.sign({sub: createdUser.id}, secret)
+    const token = jwt.sign({ sub: createdUser.id }, secret);
 
-        return res.status(201).json({ user: createdUser, token: token });
-    
+    return res.status(201).json({ user: createdUser, token: token });
 };
 
-
 const getUsers = async (req, res) => {
-    const foundUsers = await getUsersDb()
-    return res.status(200).json({users: foundUsers})
-}
-
+    const foundUsers = await getUsersDb();
+    return res.status(200).json({ users: foundUsers });
+};
 
 const deleteUser = async (req, res) => {
-    const id  = Number(req.params.id)
-    const deletedUser = await deleteUserDb(id)
-    return res.status(200).json({user: deletedUser})
-}
+    const currentUserId = Number(req.user.id);
+    const id = Number(req.params.id);
 
+    if (id === currentUserId || req.user.role === "ADMIN") {
+        const deletedUser = await deleteUserDb(id);
+        return res.status(200).json({ user: deletedUser });
+    }
+
+    return res.status(403).json({ error: "Forbidden" });
+};
 
 module.exports = {
     createUser,
     getUsers,
-    deleteUser
+    deleteUser,
 };
